\name{eco_time_audit}
\alias{eco_time_audit}
\title{Audit a time index}
\usage{eco_time_audit(data, time)}
\arguments{\item{data}{A data.frame.}\item{time}{Name of the time-index variable.}}
\value{A one-row data.frame describing the temporal index, frequency, duplicates, missingness, sorting and regularity.}
\description{Audits an explicit temporal index before explanatory time-series econometric modelling.}
